Remove the old profile list route
